feat: Add .cartesi to .dockerignore and .gitignore
Add .cartesi to the .dockerignore and .gitignore files to exclude it from version control. This ensures that the .cartesi directory is not tracked and committed.

Also, update the backend Dockerfile to include a new build stage for the backend-rs package. This stage builds the wasm package using wasm-pack and copies it to the final image.

Lastly, update the backend package.json file to use the exact version of the "@calindra/cartesify-backend" dependency.

These changes improve the project's build process and ensure that unnecessary files and directories are ignored.